[feature/v1-024] Fix minor wrongs
